Split out client code into modules.
Moved code for the different aspects of the client into
modules, each with control their own behavior and callbacks.
Removed global functions.
