+++ index.html
@@ -80,7 +80,7 @@
             <div id="form_listing" style='margin-top:20px'>
             </div>
             <h5 data-translate="title.new_form">Add New Form</h5>
-            <button onclick="javascript:selectForm();" class="btn btn-small btn-primary"><span style='margin-right:5px' data-translate="button.select_file">Select from File</span><i class="fa fa-file-o" aria-hidden="true"></i></button>
+            <button id="formChooserBtn" class="btn btn-small btn-primary"><span style='margin-right:5px' data-translate="button.select_file">Select from File</span><i class="fa fa-file-o" aria-hidden="true"></i></button>
           </div>
           <h4><i class="fa fa-map fa-lg" aria-hidden="true"></i> &#160; <span data-translate="title.tiles"></span></h4>
           <div class="well">
@@ -101,7 +101,7 @@
             <div id="track_listing" style='margin-top:20px'>
             </div>
             <h5 data-translate="title.new_track">Add New Track</h5>
-            <button onclick="javascript:selectTrackData();" class="btn btn-small btn-primary"><span style='margin-right:5px' data-translate="button.select_file">Select from File</span><i class="fa fa-file-o" aria-hidden="true"></i></button>
+            <button id="trackDataChooserBtn" class="btn btn-small btn-primary"><span style='margin-right:5px' data-translate="button.select_file">Select from File</span><i class="fa fa-file-o" aria-hidden="true"></i></button>
           </div>
         </div>
         <div role="tabpanel" class="tab-pane" id="web">
@@ -158,21 +158,21 @@
                 <span data-translate="prompt.last_synchronized">Last Synchronized: </span>
                 <span id="community_lands_sync_date" style="font-weight: bold">Unavailable</span>
               </div>
-              <div class="btn btn-small btn-primary" style="margin-top: 10px" onclick="javascript:communityLandsUpload()"><span style='margin-right:5px' data-translate="button.upload">Upload</span><i class="fa fa-cloud-upload" aria-hidden="true"></i></div>
+              <div id="communityLandsUploadBtn" class="btn btn-small btn-primary" style="margin-top: 10px"><span style='margin-right:5px' data-translate="button.upload">Upload</span><i class="fa fa-cloud-upload" aria-hidden="true"></i></div>
               <div style="margin-top: 10px" id="community_lands_sync_status"></div>
             </div>
             <h4><i class="fa fa-cloud-upload fa-lg" aria-hidden="true"></i> &#160; <span data-translate="title.data_backup">Data Backup</span></h4>
             <div class="well">
               <div id="last_backup"></div>
-              <div class="btn btn-small btn-primary" style="margin-top: 10px" onclick="javascript:backupFiles()"><span style='margin-right:5px' data-translate="button.backup_data">Backup Data</span><i class="fa fa-cloud-upload" aria-hidden="true"></i></div>
+              <div id="backupSimpleBtn" class="btn btn-small btn-primary" style="margin-top: 10px"><span style='margin-right:5px' data-translate="button.backup_data">Backup Data</span><i class="fa fa-cloud-upload" aria-hidden="true"></i></div>
               <div data-translate="help.backup.simple" class="help-block"></div>
-              <div class="btn btn-small btn-primary" style="margin-top: 10px" onclick="javascript:backupFiles(true)"><span style="margin-right:5px" data-translate="button.backup_data_usb">Backup and save to external drive</span><i class="fa fa-hdd-o" aria-hidden="true"></i></div>
+              <div id="backupUSBBtn" class="btn btn-small btn-primary" style="margin-top: 10px"><span style="margin-right:5px" data-translate="button.backup_data_usb">Backup and save to external drive</span><i class="fa fa-hdd-o" aria-hidden="true"></i></div>
               <div data-translate="help.backup.usb" class="help-block"></div>
               <div style="margin-top: 10px" id="backup_status"></div>
             </div>
             <h4><i class="fa fa-cloud-download fa-lg" aria-hidden="true"></i> &#160; <span data-translate="title.data_import">Data Import</span></h4>
             <div class="well">
-              <div class="btn btn-small btn-primary" style="margin-top: 10px" onclick="javascript:importFilesMerge()"><span style='margin-right:5px' data-translate="button.import_data_merge">Import and Merge Data</span><i class="fa fa-cloud-download" aria-hidden="true"></i></div>
+              <div id="importFilesMergeBtn" class="btn btn-small btn-primary" style="margin-top: 10px"><span style='margin-right:5px' data-translate="button.import_data_merge">Import and Merge Data</span><i class="fa fa-cloud-download" aria-hidden="true"></i></div>
               <div data-translate="help.import.merge" class="help-block"></div>
             </div>
           </div>
@@ -183,12 +183,12 @@
           <div id="settings_form">
 
           </div>
-          <button onclick="javascript:saveSettings();" class="btn btn-small btn-primary"><span style='margin-right:5px' data-translate="button.update_settings">Update Settings</span><i class="fa fa-gear" aria-hidden="true"></i></button>
+          <button id="saveSettingsBtn" class="btn btn-small btn-primary"><span style='margin-right:5px' data-translate="button.update_settings">Update Settings</span><i class="fa fa-gear" aria-hidden="true"></i></button>
           <div id="settings_status"></div>
           <div class="clearfix">&nbsp;</div>
           <h4><span data-translate="title.data_reset">Data Reset</span></h4>
           <div class="well">
-            <div class="btn btn-small btn-danger" style="margin-top: 10px" onclick="javascript:importFilesOverwrite()"><span style='margin-right:5px' data-translate="button.import_data_overwrite">Import and Overwrite Data</span><i class="fa fa-cloud-download" aria-hidden="true"></i></div>
+            <div id="importFilesOverwriteBtn" class="btn btn-small btn-danger" style="margin-top: 10px"><span style='margin-right:5px' data-translate="button.import_data_overwrite">Import and Overwrite Data</span><i class="fa fa-cloud-download" aria-hidden="true"></i></div>
             <div data-translate="help.import.overwrite" class="help-block"></div>
           </div>
